% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelProteomicsImporter.R
\name{conditionToYAMLreadyList,ExcelProteomicsImporter-method}
\alias{conditionToYAMLreadyList,ExcelProteomicsImporter-method}
\alias{conditionToYAMLreadyList}
\title{conditionToYAMLreadyList}
\usage{
\S4method{conditionToYAMLreadyList}{ExcelProteomicsImporter}(obj, colname)
}
\description{
conditionToYAMLreadyList
}
