% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelProteomicsImporter.R
\docType{class}
\name{ExcelProteomicsImporter-class}
\alias{ExcelProteomicsImporter-class}
\alias{.ExcelProteomicsImporter}
\alias{ExcelProteomicsImporter}
\title{ExcelProteomicsImporter}
\usage{
ExcelProteomicsImporter(tabDelimitedFilename)
}
\arguments{
\item{tabDelimitedFilename}{character}
}
\value{
An object of the ExcelProteomicsImporter class
}
\description{
ExcelProteomicsImporter

Create a ExcelProteomicsImporter connection
}
