% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelProteomicsImporter.R
\name{getConditionColnames,ExcelProteomicsImporter-method}
\alias{getConditionColnames,ExcelProteomicsImporter-method}
\alias{getConditionColnames}
\title{getConditionColnames}
\usage{
\S4method{getConditionColnames}{ExcelProteomicsImporter}(obj)
}
\description{
getConditionColnames
}
