% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelProteomicsImporter.R
\name{getGroupBoundaries,ExcelProteomicsImporter-method}
\alias{getGroupBoundaries,ExcelProteomicsImporter-method}
\alias{getGroupBoundaries}
\title{get groupGroupBoundaries, the start and stop of columns for avg group/s, sd group/s}
\usage{
\S4method{getGroupBoundaries}{ExcelProteomicsImporter}(obj)
}
\description{
get groupGroupBoundaries, the start and stop of columns for avg group/s, sd group/s
}
