% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelProteomicsImporter.R
\name{getGroupCount,ExcelProteomicsImporter-method}
\alias{getGroupCount,ExcelProteomicsImporter-method}
\alias{getGroupCount}
\title{get groupCount, the number of logically separate condition groups in the spreadsheet}
\usage{
\S4method{getGroupCount}{ExcelProteomicsImporter}(obj)
}
\description{
get groupCount, the number of logically separate condition groups in the spreadsheet
}
