% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelProteomicsImporter.R
\name{getRawTable,ExcelProteomicsImporter-method}
\alias{getRawTable,ExcelProteomicsImporter-method}
\alias{getRawTable}
\title{get the original, raw data.frame}
\usage{
\S4method{getRawTable}{ExcelProteomicsImporter}(obj)
}
\description{
get the original, raw data.frame
}
