% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelProteomicsImporter.R
\name{identifyGroupBoundaries,ExcelProteomicsImporter-method}
\alias{identifyGroupBoundaries,ExcelProteomicsImporter-method}
\alias{identifyGroupBoundaries}
\title{identifyGroupBoundaries: where avg and sd columns begin and end}
\usage{
\S4method{identifyGroupBoundaries}{ExcelProteomicsImporter}(obj)
}
\description{
identifyGroupBoundaries: where avg and sd columns begin and end
}
