% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelProteomicsImporter.R
\name{toYAML,ExcelProteomicsImporter-method}
\alias{toYAML,ExcelProteomicsImporter-method}
\alias{toYAML}
\title{get the matrix of averages}
\usage{
\S4method{toYAML}{ExcelProteomicsImporter}(obj)
}
\description{
get the matrix of averages
}
