% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelProteomicsImporter.R
\name{validateColumnNames,ExcelProteomicsImporter-method}
\alias{validateColumnNames,ExcelProteomicsImporter-method}
\alias{validateColumnNames}
\title{learn experiment conditions and timepoints from the column names}
\usage{
\S4method{validateColumnNames}{ExcelProteomicsImporter}(obj)
}
\description{
learn experiment conditions and timepoints from the column names
}
